Revert "Acpica trivial"

+++ source/compiler/aslascii.c
@@ -252,8 +252,8 @@ FlIsFileAsciiSource (
         if (DisplayErrors)
         {
             AcpiOsPrintf (
-                "Total %lu invalid characters found in input source text, "
-                "could be a binary file\n", (UINT64) BadBytes);
+                "Total %u invalid characters found in input source text, "
+                "could be a binary file\n", BadBytes);
             AslError (ASL_ERROR, ASL_MSG_NON_ASCII, NULL, Filename);
         }
